requirements.txt
click==6.7
ecdsa==0.13
Fabric3==1.12.post1
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
requests==2.12.4
six==1.10.0
Werkzeug==0.11.15
